% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_access_token.R
\name{create_access_token}
\alias{create_access_token}
\title{Generate access token}
\usage{
create_access_token(consumer_key, consumer_secret_key)
}
\arguments{
\item{consumer_key}{Alphanumeric consumer key provided by OPS}

\item{consumer_secret_key}{Alphanumeric secret provided by OPS}
}
\value{
Access token in a string format
}
\description{
Generation of access token for OPS authentication. The access token expires every 20 minutes.
}
\examples{
\dontrun{create_access_token(consumer_key, consumer_secret_key)}


}
