% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffs.R
\docType{methods}
\name{diffs}
\alias{diffs}
\alias{diffs,list,NULL-method}
\alias{diffs,NULL,list-method}
\alias{diffs,ANY,NULL-method}
\alias{diffs,NULL,ANY-method}
\alias{diffs,atomic,atomic-method}
\alias{diffs,list,list-method}
\title{Print differences between two objects}
\usage{
diffs(x, y, .depth = 0L, .name = NULL, .type = 0L)

\S4method{diffs}{list,`NULL`}(x, y, .depth = 0L, .name = NULL, .type = 0L)

\S4method{diffs}{`NULL`,list}(x, y, .depth = 0L, .name = NULL, .type = 0L)

\S4method{diffs}{ANY,`NULL`}(x, y, .depth = 0L, .name = NULL, .type = 0L)

\S4method{diffs}{`NULL`,ANY}(x, y, .depth = 0L, .name = NULL, .type = 0L)

\S4method{diffs}{atomic,atomic}(x, y, .depth = 0L, .name = NULL, .type = 0L)

\S4method{diffs}{list,list}(x, y, .depth = 0L, .name = NULL, .type = 0L)
}
\arguments{
\item{x}{(\code{list}, \code{atomic}) object to compare.}

\item{y}{(\code{list}, \code{atomic}) obejct to compare with.}

\item{.depth}{(\code{integer(1)}) Private argument, please don't specify.
depth of the call. Increases by \code{1} according to depth of the lists \code{x} and \code{y}}

\item{.name}{(\code{character(1)}) Private argument, please don't specify.
Used to identify the name of the current list element when calld resursivelly.
\code{.name} can be used with \code{.expr}.}

\item{.type}{(\code{integer(1)}) Private argument, please don't specify.
\code{.type} denotes what's the result of comparison for this element in the \code{parent.call}.
\itemize{
\item \code{0} if \code{x} and the \code{y} exist in the recursive path.
\item \code{1} if only \code{x} exists in the recursive path.
\item \code{2} if only \code{y} exists in the recursive path.
}}
}
\value{
invisible \code{NULL}
}
\description{
Print differences between two objects.
}
\examples{
diffs(c(1, 2), c(2, 3))
diffs(c(a = 1, b = 2), c(c = 2, d = 3))
diffs(letters[1:2], letters[2:3])
}
