% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen-params.R
\name{.params}
\alias{.params}
\title{Parameters only}
\usage{
.params(x, y, .expr, .name, .depth, .type)
}
\arguments{
\item{x}{(\code{list}, \code{atomic}) object to compare.}

\item{y}{(\code{list}, \code{atomic}) obejct to compare with.}

\item{.expr}{(\code{language}) any syntactically valid R expression.
Argument is parsed, evaluated and passed recursivelly to calls run on the nested
elements of \code{x} and \code{y}. Other arguments (\code{.name}, \code{.depth} and \code{.type}) can be used
to influence output of the comparisons.
For example \code{if (.depth == 2) return(x)} will return \code{x} for the nested list elements
of depth equal \code{2}.
Note: If \code{.expr} has a \code{return} statement please be aware that it should be the
same as the type of the output.}

\item{.name}{(\code{character(1)}) Private argument, please don't specify.
Used to identify the name of the current list element when calld resursivelly.
\code{.name} can be used with \code{.expr}.}

\item{.depth}{(\code{integer(1)}) Private argument, please don't specify.
depth of the call. Increases by \code{1} according to depth of the lists \code{x} and \code{y}}

\item{.type}{(\code{integer(1)}) Private argument, please don't specify.
\code{.type} denotes what's the result of comparison for this element in the \code{parent.call}.
\itemize{
\item \code{0} if \code{x} and the \code{y} exist in the recursive path.
\item \code{1} if only \code{x} exists in the recursive path.
\item \code{2} if only \code{y} exists in the recursive path.
}}
}
\description{
Parameters to inherit from
}
\keyword{internal}
