% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_abstracts.R
\name{extract_abstracts}
\alias{extract_abstracts}
\title{Extract abstracts during bibliographic patent data parsing}
\usage{
extract_abstracts(x)
}
\arguments{
\item{x}{abstract list}
}
\value{
list containing abstracts
}
\description{

}
