% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_IPC_class.R
\name{extract_IPC_class}
\alias{extract_IPC_class}
\title{Extract IPC classification during patent data parsing}
\usage{
extract_IPC_class(x)
}
\arguments{
\item{x}{IPC location}
}
\value{
vector containing IPC classification of various patents
}
\description{

}
