% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_titles.R
\name{extract_titles}
\alias{extract_titles}
\title{Extract titles during bibliographic patent data parsing}
\usage{
extract_titles(x)
}
\arguments{
\item{x}{title list}
}
\value{
list containing titles
}
\description{

}
