% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstract.R
\name{get_abstract}
\alias{get_abstract}
\title{Get patent abstract}
\usage{
get_abstract(id, type, format, access_token, fromRange = 1, toRange = 100)
}
\arguments{
\item{type}{publication or application (pub or app)}

\item{format}{id type of the patent, it must be either epodoc or docdb}

\item{access_token}{access token generated using the create_access_token function}

\item{patent_id}{identification code of the patent. It can either be in the epodoc or docdb format}
}
\value{
patent bibliographic information as dataframe
}
\description{
Retrieves up to 100 patent abstracts through the epodoc or docdb id
}
\examples{
\dontrun{search_patent(docdb_id, access_token))}


}
