% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biblio.R
\name{get_biblio}
\alias{get_biblio}
\title{Get patent bibliographic data}
\usage{
get_biblio(docdb_id, access_token)
}
\arguments{
\item{docdb_id}{docdb identification code of the patent}

\item{access_token}{access token generated using the create_access_token function}
}
\value{
patent bibliographic information in json format
}
\description{
search for a patent
}
\examples{
\dontrun{search_patent(docdb_id, access_token))}


}
