% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_claims.R
\name{get_claims}
\alias{get_claims}
\title{Get claims}
\usage{
get_claims(epodoc, access_token)
}
\arguments{
\item{epodoc}{epodoc code for a patent to have returned the claims}

\item{access_token}{token for authentications}
}
\value{
claims contained in the patent as a list
}
\description{
Retrieve claims of a patent. It works only with the epodoc format
}
\examples{
\dontrun{get_claims(epodoc, access_token))}


}
