% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\title{Get description}
\usage{
get_description(epodoc, access_token)
}
\arguments{
\item{epodoc}{epodoc identification code}

\item{access_token}{token for authentication}
}
\value{
despcription contained in the patent
}
\description{
Retrieve patent description based on epodoc identification code
}
\examples{
\dontrun{get_description(epodoc, access_token))}


}
