% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ops.R
\name{get_ops}
\alias{get_ops}
\title{Obtain a set of urls from ops_publications}
\usage{
get_ops(url, access_token, raw = NULL)
}
\arguments{
\item{url}{. A single url or vector of urls}

\item{access_token}{bearer token for authentication}
}
\value{
A list.
}
\description{
Used in conjunction with ops_iterate. Given a set of urls convert to a
get request to retrieve the data from OPS. When the response is received extract the content to create a list.
}
\examples{
\dontrun{lapply(three_urls, ops_get)}
}
