% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request.R
\name{get_request}
\alias{get_request}
\title{Makes a get request to the Open Patent Services API}
\usage{
get_request(url, access_token, raw = FALSE)
}
\arguments{
\item{url}{string, A single url}

\item{access_token}{string, bearer token for authentication}

\item{raw}{bool, r}
}
\value{
A list
}
\description{
Makes a request to the Open Patent Services API
and it returns the raw or parsed response
}
\examples{
\dontrun{
response <- get_request(url, access_token, raw = FALSE)
}

}
