% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result_number.R
\name{get_result_number}
\alias{get_result_number}
\title{Retrieve number of result matching a query created using get_query}
\usage{
get_result_number(query, access_token)
}
\arguments{
\item{query}{url built using the function get_query}

\item{access_token}{access token received after authentication}
}
\value{

}
\description{
This is a memory efficient way to count the number of results of a single query. Whenever it returns 10000, it means that the number of results are 10000 or higher
}
\examples{
\dontrun{get_result_number(query, access_token)}

}
