% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescing-ops.R
\name{ifna}
\alias{ifna}
\title{Missingness Coalescing Operator}
\usage{
ifna(x, y)
}
\arguments{
\item{x}{An possibly containing \code{NA} values}

\item{y}{An object of equal length or a single value to be used in
substitution}
}
\description{
Substitutes values when \code{NA} is detected.
}
\details{
The objective of this function is to provide an ability to substitute values
for NA. However, it is important to note that \emph{R} is unique in the fact
that it has an \code{NA} data type to represent missingness instead of relying
on values being pre-coded (e.g. \code{0}, \code{-1111}, et cetera). Thus, the simplicity
of this function comes with the disclaimer of:

"An NA is the presence of an absence.
Don't forget that some missing values are the absence of a presence"
\itemize{
\item Hadley Wickham on Twitter
}
}
\examples{
# Data with missing values
x <- c(1, NA, NA, 4)
# Substitution vector of equal length
y <- 1:4

# Replace NA with values in `y` vector
ifna(x, y)

# Replace NA with 5
ifna(x, 5)
}
\references{
\url{https://support.office.com/en-us/article/IFNA-function-6626c961-a569-42fc-a49d-79b4951fd461}
}
