% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalescing-ops.R
\name{\%??\%}
\alias{\%??\%}
\title{Null Coalescing Operator}
\usage{
x \%??\% y
}
\arguments{
\item{x}{A value that is possible \code{NULL}}

\item{y}{A value that is not \code{NULL}}
}
\value{
Returns the value of the left hand side (LHS) if it is not null, else
returns the value of the right hand side (RHS) that may be \code{NULL}.
}
\description{
Binary operator to simplify null expressions
}
\details{
The objective of the null coalescing operator is to simplify the expression:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{if(is.null(x)) \{
  y
else \{
  x
\}
}\if{html}{\out{</div>}}

This allows it to be readable inline.
}
\section{Warning}{

Due to the way objects are created within \emph{R}, the \code{NULL} value
is not able to be stored within \code{\link[=vector]{atomic vectors}}.
However, the \code{\link{NULL}} value can be stored within a \code{\link{list}}.
This operator will \emph{not} be triggered if \code{\link{NULL}} is within
a list! That is, the RHS side will be returned.
}

\examples{
# Null value
x <- NULL

# Before
y <- if(is.null(x)){ "Unset" } else { x }

# After
y <- x \%??\% "Unset"

# Concrete examples without variables
# Returns 3 as the LHS is _not_ NULL
3 \%??\% 4

# Returns 1 as the LHS is _not_ NULL
1 \%??\% NULL

# Returns 2 as the LHS is NULL
NULL \%??\% 2

# Coalesce operator
NULL \%??\% NULL \%??\% 5

# Coalesce operator
NULL \%??\% 7 \%??\% 8
}
\references{
\url{https://en.wikipedia.org/wiki/Null_coalescing_operator}
}
\seealso{
\code{\link[=is.null]{is.null()}}, \code{\link{NULL}}
}
