% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_biblio.R
\name{parse_biblio}
\alias{parse_biblio}
\title{Parse json data returned from the biblio endpoint into dataframe}
\usage{
parse_biblio(response)
}
\arguments{
\item{response}{A single url or vector of urls}
}
\value{
A list.
}
\description{
Given a json list after fromJSON() function of the jsonlite package, it returns a dataframe conteining some bibliographic patent data in an orderly manner. Tt works only with results in multiple entries. Not stable, as the OPS is still developing, this method needs to be improved.
}
\examples{
\dontrun{parse_biblio(response)}
}
