% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_search.R
\name{parse_search}
\alias{parse_search}
\title{Parse json response into docdb codes of patents}
\usage{
parse_search(response)
}
\arguments{
\item{response}{access token generated using the create_access_token function}
}
\value{
character vector containing docdb codes of query response
}
\description{
Parse json response into docdb codes of patents
}
\examples{
\dontrun{parse_search(response)}
}
