% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_services.R
\name{publication_services}
\alias{publication_services}
\title{Queries the Publication Services Open Patent Services API}
\usage{
publication_services(
  id,
  type,
  format,
  access_token,
  what = "fulltext",
  raw = FALSE
)
}
\arguments{
\item{id}{string or string vector, epodoc or docdb format}

\item{type}{the type of publication to be searched.
'publication' or 'application'.}

\item{format}{the format of the document id to be searched.
'epodoc' or 'docdb'.}

\item{access_token}{token for authentications}

\item{what}{string, the type of information to be retrieved.
'claims', 'abstract', 'description', 'title', 'fulltext', 'full-cycle'}

\item{raw}{bool, if TRUE returns the raw response}
}
\value{
A list or a raw string
}
\description{
Retrieve claims of a patent.
It works only with the epodoc format
}
\examples{
\dontrun{publication_services(id = "EP1000000.A1", type = "publication", format = "epodoc", access_token = "your_access_token", raw = FALSE, what = "claims")}
}
