% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rops-package.R
\docType{package}
\name{rops-package}
\alias{rops}
\alias{rops-package}
\title{rops: Set operations for lists and its elements}
\description{
Recursive union, intersect, setdiff for the lists and its elements.
}
\author{
\strong{Maintainer}: Dawid Kałędkowski \email{dawid.kaledkowski@gmail.com} (\href{https://orcid.org/0000-0001-9533-457X}{ORCID})

}
\keyword{internal}
