% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_patents.R
\name{search_patents}
\alias{search_patents}
\title{Search patents}
\usage{
search_patents(
  title = NA,
  abstract = NA,
  titab = NA,
  applicant = NA,
  inventor = NA,
  citation = NA,
  pub_num = NA,
  cpc = NA,
  ipc = NA,
  from = NA,
  to = NA,
  with_biblio = FALSE,
  access_token,
  fromRange = 1,
  toRange = 100
)
}
\arguments{
\item{access_token}{access token received after authentication}

\item{query}{url built using the function get_query}
}
\value{
Dataframe containing patents
}
\description{
Search for patents contained in the EPO dataset
}
\examples{
\dontrun{search_patents(query, access_token)}
}
