% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biblio.R
\name{get_biblio}
\alias{get_biblio}
\title{Get patent bibliographic data}
\usage{

get_biblio(id, type, format, access_token, fromRange, toRange)
}
\arguments{
\item{access_token}{access token generated using the create_access_token function}

\item{patent_id}{identification code of the patent. It can either be in the epodoc or docdb format}

\item{id_type}{id type of the patent, it must be either epodoc or docdb}
}
\value{
patent bibliographic information in json format
}
\description{
search for a patent
}
\examples{
\dontrun{search_patent(docdb_id, access_token))}


}
