% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ops.R
\name{get_ops}
\alias{get_ops}
\title{Makes a request to the Open Patent Services API}
\usage{
get_ops(url, access_token, raw = NULL)
}
\arguments{
\item{url}{string, A single url}

\item{access_token}{string, bearer token for authentication}

\item{raw}{bool, r}
}
\value{
A list
}
\description{
Makes a request to the Open Patent Services API
and it returns the raw or parsed response
}
\examples{
\dontrun{
search_respose <- get_ops(url, access_token, raw = FALSE)
}

}
