% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ops.R
\name{get_ops}
\alias{get_ops}
\title{Obtain a set of urls from ops_publications}
\usage{
get_ops(url, access_token, raw = NULL, from_range = 1, to_range = 100)
}
\arguments{
\item{url}{. A single url}

\item{access_token}{bearer token for authentication}

\item{raw}{boolean value that decides whether the returned file will be in json format or as a list}

\item{from_range}{Initial limit of results to be returned. Default to 1}

\item{to_range}{Final limit of results to be returned. Default to 100}
}
\value{
A list containing a parse
}
\description{
Used in conjunction with ops_iterate. Given a set of urls convert to a
get request to retrieve the data from OPS. When the response is received extract the content to create a list.
}
\examples{
\dontrun{search_respose <- get_ops("https://ops.epo.org/3.2/rest-services/published-data/search/?q=ct\%3D CN104980769 \%20 and \%20 pd \%3D \%22 20180201\%2020180203\%22", access_token, raw = FALSE)}
}
