% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_urls.R, R/get_query.R
\name{get_query}
\alias{get_query}
\title{Generate URLs to search OPS biblios}
\usage{
get_query(
  title = NA,
  abstract = NA,
  titab = NA,
  applicant = NA,
  inventor = NA,
  citation = NA,
  pub_num = NA,
  cpc = NA,
  ipc = NA,
  from = NA,
  to = NA
)

get_query(
  title = NA,
  abstract = NA,
  titab = NA,
  applicant = NA,
  inventor = NA,
  citation = NA,
  pub_num = NA,
  cpc = NA,
  ipc = NA,
  from = NA,
  to = NA
)
}
\arguments{
\item{title}{- character}

\item{abstract}{- character}

\item{applicant}{- character}

\item{inventor}{- character}

\item{citation}{- character}

\item{cpc}{- character}

\item{ipc}{- character}

\item{from}{- YYYY or YYYYMMDD, publication date.}

\item{to}{- YYYY or YYYYMMDD, publication date}
}
\value{
prints a query that is ready to be used in the Publication Data Keyword Search endpoint

prints a query that is ready to be used in the Publication Data Keyword Search endpoint
}
\description{
Generate URLs to search OPS biblios

Generate URLs to search OPS biblios
}
\details{
This function generates a set of URLs with a maximum of 100 results per URL for searching either OPS titles, titles & abstracts, biblios (default) and/or date ranges (publication date). See Details.

This function generates a set of URLs with a maximum of 100 results per URL for searching either OPS titles, titles & abstracts, biblios (default) and/or date ranges (publication date). See Details.
}
\examples{
\dontrun{urls <- get_query(title = "pizza", applicant = "IBM", from = 2010, to = 2011)}
\dontrun{urls <- get_quert(title = "pizza", applicant = "IBM", from = 2010, to = 2011)}
}
