% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality-ops.R
\name{\%??\%}
\alias{\%??\%}
\alias{ifnull}
\title{Null Coalescing Operator}
\usage{
x \%??\% y

ifnull(x, y)
}
\arguments{
\item{x}{A value that is possible \code{NULL}}

\item{y}{A value that is not \code{NULL}}
}
\value{
Returns the value of the left hand side (LHS) if it is not null, else
returns the value of the right hand side (RHS) that may be \code{NULL}.
}
\description{
Binary operator to simplify null expressions
}
\details{
The objective of the null coalescing operator is to simplify the expression:
\code{if(is.null(x)){ y } else { x }} so that it is readable inline.
}
\examples{
# Returns 3 as the LHS is _not_ NULL
3 \%??\% 4

# Returns 1 as the LHS is _not_ NULL
1 \%??\% NULL

# Returns 2 as the LHS is NULL
NULL \%??\% 2

# Coalesce operator
NULL \%??\% NULL \%??\% 5

# Coalesce operator
NULL \%??\% 7 \%??\% 8
}
\references{
\url{https://en.wikipedia.org/wiki/Null_coalescing_operator}
}
\seealso{
\code{\link{is.null}}
}

