% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_search.R
\name{parse_search}
\alias{parse_search}
\title{Obtain a set of urls from ops_publications}
\usage{
parse_search(response)
}
\arguments{
\item{response}{json list returned from get_ops function}

\item{url}{. A single url or vector of urls}
}
\value{
A list.
}
\description{
Used in conjunction with ops_iterate. Given a set of urls convert to a
get request to retrieve the data from OPS. When the response is received extract the content to create a list.
}
\examples{
\dontrun{lapply(three_urls, ops_get)}

}
