% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_search_biblio.R
\name{parse_search_biblio}
\alias{parse_search_biblio}
\title{Parse search response with attached biblio}
\usage{
parse_search_biblio(response)
}
\arguments{
\item{response}{json list returned from get_ops function}
}
\value{
A dataframe containing the search results
}
\description{
This function transforms the response obtained by the GET request to the OPS into a dataframe containing the search results.
}
\examples{
\dontrun{parse_search_biblio(response)}

}
