% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_request.R
\name{post_request}
\alias{post_request}
\title{Makes a POST request to the Open Patent Services API}
\usage{
post_request(url, access_token, raw = raw, id = NULL)
}
\arguments{
\item{url}{string, A single url}

\item{access_token}{string, bearer token for authentication}

\item{raw}{bool, r}
}
\value{
A list
}
\description{
Makes a POST request to the Open Patent Services API
and it returns the raw or parsed response
}
\examples{
\dontrun{
search_respose <- get_ops(url, access_token, raw = FALSE)
}

}
