% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_services.R
\name{publication_services}
\alias{publication_services}
\title{Queries the Publication Services Open Patent Services API}
\usage{
publication_services(
  id,
  type,
  format,
  access_token,
  what = "fulltext",
  raw = FALSE,
  equilvalents = NULL
)
}
\arguments{
\item{type}{the type of publication to be searched. 'pub' stands for patent publications and 'app' stands for patent applications.}

\item{access_token}{token for authentications}

\item{epodoc_id}{character, epodoc code for a patent to have returned the claims}
}
\value{
A character vector containing patent claims
}
\description{
Retrieve claims of a patent. It works only with the epodoc format
}
\examples{
\dontrun{get_full_text(epodoc_id, type = 'pub', access_token))}
}
